+++ GDX.Tests/EditMode/Threading/TaskDirectorTests.cs
@@ -11,5 +11,29 @@ namespace GDX.Threading
     /// </summary>
     public class TaskDirectorTests
     {
+        // [Test]
+        // [Category(Core.TestCategory)]
+        // public void TaskDirector_Logging()
+        // {
+        //     new TestTask(TestLiterals.HelloWorld, 5, null, null).Enqueue();
+        //
+        //     TaskDirector.Complete();
+        //
+        //     Assert.IsTrue(m_Log.Count == 5, $"{m_Log.Count} != 5");
+        // }
+
+        // OnBlockUserInput
+        // OnLogAdded
+
+        // AddLog
+        // GetBusyCount
+        // GetQueueCount
+        // GetStatus
+        // HasTasks
+        // QueueTask
+        // Tick
+        // UpdateTask
+        // Wait
+        // WaitFor/
     }
 }