+++ CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ***"Dinner Table"***
 >
 
-### Addded
+### Added
 
 - New toggle added to Project Settings -> GDX -> Environment, ensuring that shaders included are always included in a build
-- 
+- `TaskDirector` to assist with `TaskBase` based scheduling of the thread pool.
+- `EditorTaskDirector` to automatically tick the `TaskDirector` while in edit mode.
 
 ### Fixed